refactor(responses): separate native and normalized executors
